Database added and connected
Goal is to have the game create rooms for users to create and join.
A lobby component has been added for the user to choose from creating a room or joining a room. When creating a room, the user sends the same input from the last version to the server. The server adds this information to the database and returns the room information, with one user; which the player state is created from. The playerslice has been updated to receive this new object formed inside thunks.js. I wanted to have a more complete picture completed before commiting, however it's not good practice and will do smaller commits from now on.