Join existing game functionality has been added
If an id is passed to the fetchroom thunk, then the passed user info will update an existing player or add a new player to the room. Minor fixes made throughout. New RoomInfo component created to handle the selection of a player to an existing room. An API route for room/:id was added to update existing room and send back the room with the user attached.