Ball locations and whoseturn gets fupdated in db
Updated the method of updating whoseturn to be updated in database first then sending the update through websockets to connected players. This way when a player reconnects to room they will have the latest information, same for ball locations. Updated the websocket emission of ending turn to send who goes next to each connected player, instead of having the reducer check it's state then update the player locally. Updated name to username throughout the app to separate the 'name' property for room name in database.